Share ISyntaxFormattingService and implementations
