+++ src/Workspaces/Core/Portable/Formatting/AbstractSyntaxFormattingService.cs
@@ -4,7 +4,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
-using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.Formatting.Rules;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.Shared;
@@ -28,9 +27,9 @@ namespace Microsoft.CodeAnalysis.Formatting
 
         protected abstract IFormattingResult CreateAggregatedFormattingResult(SyntaxNode node, IList<AbstractFormattingResult> results, SimpleIntervalTree<TextSpan> formattingSpans = null);
 
-        protected abstract Task<AbstractFormattingResult> FormatAsync(SyntaxNode node, OptionSet options, IEnumerable<IFormattingRule> rules, SyntaxToken token1, SyntaxToken token2, CancellationToken cancellationToken);
+        protected abstract AbstractFormattingResult Format(SyntaxNode node, OptionSet options, IEnumerable<IFormattingRule> rules, SyntaxToken token1, SyntaxToken token2, CancellationToken cancellationToken);
 
-        public Task<IFormattingResult> FormatAsync(SyntaxNode node, IEnumerable<TextSpan> spans, OptionSet options, IEnumerable<IFormattingRule> rules, CancellationToken cancellationToken)
+        public IFormattingResult Format(SyntaxNode node, IEnumerable<TextSpan> spans, OptionSet options, IEnumerable<IFormattingRule> rules, CancellationToken cancellationToken)
         {
             CheckArguments(node, spans, options, rules);
 
@@ -38,19 +37,19 @@ namespace Microsoft.CodeAnalysis.Formatting
             var spansToFormat = new NormalizedTextSpanCollection(spans.Where(s_notEmpty));
             if (spansToFormat.Count == 0)
             {
-                return Task.FromResult(CreateAggregatedFormattingResult(node, SpecializedCollections.EmptyList<AbstractFormattingResult>()));
+                return CreateAggregatedFormattingResult(node, SpecializedCollections.EmptyList<AbstractFormattingResult>());
             }
 
             // check what kind of formatting strategy to use
             if (AllowDisjointSpanMerging(spansToFormat, options.GetOption(FormattingOptions.AllowDisjointSpanMerging)))
             {
-                return FormatMergedSpanAsync(node, options, rules, spansToFormat, cancellationToken);
+                return FormatMergedSpan(node, options, rules, spansToFormat, cancellationToken);
             }
 
-            return FormatIndividuallyAsync(node, options, rules, spansToFormat, cancellationToken);
+            return FormatIndividually(node, options, rules, spansToFormat, cancellationToken);
         }
 
-        private async Task<IFormattingResult> FormatMergedSpanAsync(
+        private IFormattingResult FormatMergedSpan(
             SyntaxNode node, OptionSet options, IEnumerable<IFormattingRule> rules, IList<TextSpan> spansToFormat, CancellationToken cancellationToken)
         {
             var spanToFormat = TextSpan.FromBounds(spansToFormat[0].Start, spansToFormat[spansToFormat.Count - 1].End);
@@ -62,11 +61,11 @@ namespace Microsoft.CodeAnalysis.Formatting
             }
 
             // more expensive case
-            var result = await FormatAsync(node, options, rules, pair.Item1, pair.Item2, cancellationToken).ConfigureAwait(false);
+            var result = Format(node, options, rules, pair.Item1, pair.Item2, cancellationToken);
             return CreateAggregatedFormattingResult(node, new List<AbstractFormattingResult>(1) { result }, SimpleIntervalTree.Create(TextSpanIntervalIntrospector.Instance, spanToFormat));
         }
 
-        private async Task<IFormattingResult> FormatIndividuallyAsync(
+        private IFormattingResult FormatIndividually(
             SyntaxNode node, OptionSet options, IEnumerable<IFormattingRule> rules, IList<TextSpan> spansToFormat, CancellationToken cancellationToken)
         {
             List<AbstractFormattingResult> results = null;
@@ -78,7 +77,7 @@ namespace Microsoft.CodeAnalysis.Formatting
                 }
 
                 results = results ?? new List<AbstractFormattingResult>();
-                results.Add(await FormatAsync(node, options, rules, pair.Item1, pair.Item2, cancellationToken).ConfigureAwait(false));
+                results.Add(Format(node, options, rules, pair.Item1, pair.Item2, cancellationToken));
             }
 
             // quick simple case check
